feat(data-table): improve usage
